+++ tests/round_trips.rs
@@ -98,7 +98,7 @@ fn filesystem_chunk_round_trip_impl(
 #[test]
 fn chunk_round_trip_filesystem_key_encoding_default_slash() -> Result<(), Box<dyn Error>> {
     let path = tempfile::TempDir::new()?;
-    let chunk_key_encoding: ChunkKeyEncoding = Box::new(DefaultChunkKeyEncoding::default());
+    let chunk_key_encoding = ChunkKeyEncoding::new(DefaultChunkKeyEncoding::default());
     filesystem_chunk_round_trip_impl(path.path(), &chunk_key_encoding)?;
     let mut path_expect = path.path().to_owned();
     path_expect.push("group/array/c/0/0/0");
@@ -109,7 +109,7 @@ fn chunk_round_trip_filesystem_key_encoding_default_slash() -> Result<(), Box<dy
 #[test]
 fn chunk_round_trip_filesystem_key_encoding_default_dot() -> Result<(), Box<dyn Error>> {
     let path = tempfile::TempDir::new()?;
-    let chunk_key_encoding: ChunkKeyEncoding = Box::new(DefaultChunkKeyEncoding::new_dot());
+    let chunk_key_encoding = ChunkKeyEncoding::new(DefaultChunkKeyEncoding::new_dot());
     filesystem_chunk_round_trip_impl(path.path(), &chunk_key_encoding)?;
     let mut path_expect = path.path().to_owned();
     path_expect.push("group/array/c.0.0.0");
@@ -120,7 +120,7 @@ fn chunk_round_trip_filesystem_key_encoding_default_dot() -> Result<(), Box<dyn 
 #[test]
 fn chunk_round_trip_filesystem_key_encoding_v2_dot() -> Result<(), Box<dyn Error>> {
     let path = tempfile::TempDir::new()?;
-    let chunk_key_encoding: ChunkKeyEncoding = Box::new(V2ChunkKeyEncoding::default());
+    let chunk_key_encoding = ChunkKeyEncoding::new(V2ChunkKeyEncoding::default());
     filesystem_chunk_round_trip_impl(path.path(), &chunk_key_encoding)?;
     let mut path_expect = path.path().to_owned();
     path_expect.push("group/array/0.0.0");
@@ -131,7 +131,7 @@ fn chunk_round_trip_filesystem_key_encoding_v2_dot() -> Result<(), Box<dyn Error
 #[test]
 fn chunk_round_trip_filesystem_key_encoding_v2_slash() -> Result<(), Box<dyn Error>> {
     let path = tempfile::TempDir::new()?;
-    let chunk_key_encoding: ChunkKeyEncoding = Box::new(V2ChunkKeyEncoding::new_slash());
+    let chunk_key_encoding = ChunkKeyEncoding::new(V2ChunkKeyEncoding::new_slash());
     filesystem_chunk_round_trip_impl(path.path(), &chunk_key_encoding)?;
     let mut path_expect = path.path().to_owned();
     path_expect.push("group/array/0/0/0");